+++ packages/scaffold/src/partials/w3m-account-default-widget/styles.ts
@@ -86,11 +86,11 @@ export default css`
     height: 48px;
     padding: 10px;
     flex: 1 0 0;
-
     border-radius: var(--XS, 16px);
     border: 1px solid var(--dark-accent-glass-010, rgba(71, 161, 255, 0.1));
     background: var(--dark-accent-glass-010, rgba(71, 161, 255, 0.1));
-    transition: background 0.2s linear;
+    transition: background-color var(--wui-ease-out-power-1) var(--wui-duration-md);
+    will-change: background-color;
   }
 
   .account-links wui-flex:hover {