+++ packages/ui/src/utils/TypeUtil.ts
@@ -11,6 +11,7 @@ export type ColorType =
   | 'inverse-100'
   | 'success-100'
   | 'glass-005'
+  | 'glass-020'
 
 export type TextType =
   | 'large-500'
@@ -108,6 +109,7 @@ export type IconType =
   | 'browser'
   | 'card'
   | 'checkmark'
+  | 'checkmarkBold'
   | 'chevronBottom'
   | 'chevronLeft'
   | 'chevronRight'
@@ -146,6 +148,7 @@ export type IconType =
   | 'swapHorizontal'
   | 'swapHorizontalBold'
   | 'swapHorizontalMedium'
+  | 'swapHorizontalRoundedBold'
   | 'swapVertical'
   | 'telegram'
   | 'twitch'
@@ -165,7 +168,9 @@ export type VisualType =
   | 'defi'
   | 'defiAlt'
   | 'eth'
+  | 'google'
   | 'layers'
+  | 'lightbulb'
   | 'lock'
   | 'login'
   | 'network'
@@ -177,6 +182,7 @@ export type VisualType =
   | 'moonpay'
   | 'stripe'
   | 'paypal'
+  | 'pencil'
 
 export type VisualSize = 'sm' | 'md' | 'lg'